requirements.txt
flask>=2.3.0
requests>=2.31.0
python-whois>=0.8.0
urllib3>=2.0.0
